refactor(persistence): split event store internals
