+++ src/persistence/event-store.ts
@@ -1,1176 +1,32 @@
-import { createHash } from 'node:crypto';
-import { chmod, mkdir, open } from 'node:fs/promises';
-import path from 'node:path';
-
-import {
-  createUsageEvent,
-  normalizeSourceId,
-  type UsageEvent,
-  type UsageEventInput,
-} from '../domain/usage-event.js';
-import { normalizeProviderToBillingEntity } from '../domain/provider-normalization.js';
-import { normalizeSkippedRowReasons } from '../cli/normalize-skipped-row-reasons.js';
-import { loadNodeSqliteModule } from '../sources/opencode/node-sqlite-loader.js';
-import type { SourceSkippedRowReasonStat } from '../sources/source-adapter.js';
-import { asRecord } from '../utils/as-record.js';
-import { compareByCodePoint } from '../utils/compare-by-code-point.js';
-import { getUserCacheRootDir } from '../utils/cache-root-dir.js';
-
-export const EVENT_STORE_SCHEMA_VERSION = '3';
-
-// Migration rehash pages through events with keyset pagination so a large
-// store never materializes every row in memory at once.
-export const MIGRATION_BATCH_SIZE = 5_000;
-
-const EVENT_STORE_OPEN_TIMEOUT_MS = 2_000;
-const CONTROL_CHARACTERS_PATTERN = new RegExp(String.raw`[\u0000-\u001F\u007F-\u009F]`, 'u');
-const NORMALIZED_TIMESTAMP_PATTERN = /^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{3}Z$/u;
-const DAYS_IN_MONTH = [31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31] as const;
-const FAST_PATH_REJECT = Symbol('fast path reject');
-
-type EventStoreStatement = {
-  all: (...parameters: unknown[]) => Record<string, unknown>[];
-  get: (...parameters: unknown[]) => Record<string, unknown> | undefined;
-  run: (...parameters: unknown[]) => unknown;
-  setReturnArrays: (enabled: boolean) => void;
-};
-
-type EventStoreDatabase = {
-  exec: (sql: string) => void;
-  prepare: (sql: string) => EventStoreStatement;
-  close: () => void;
-};
-
-type EventStoreSqliteModule = {
-  DatabaseSync: new (
-    filePath: string,
-    options?: {
-      readOnly?: boolean;
-      timeout?: number;
-    },
-  ) => EventStoreDatabase;
-};
-
-export type LoadEventStoreSqliteModule = () => Promise<unknown>;
-
-const loadEventStoreSqliteModule: LoadEventStoreSqliteModule = () =>
-  loadNodeSqliteModule('Event store');
-
-export type EventStoreDependencyFingerprint = {
-  path: string;
-  exists: boolean;
-  size?: number;
-  mtimeMs?: number;
-};
-
-export type EventStoreFileFingerprint = {
-  dependencies: EventStoreDependencyFingerprint[];
-};
-
-export type EventStoreFileEntry = {
-  fingerprint: string;
-  skippedRows: number;
-  skippedRowReasons: SourceSkippedRowReasonStat[];
-};
-
-export type EventStore = {
-  database: EventStoreDatabase;
-  filePath: string;
-  statements: {
-    getFileEntry?: EventStoreStatement;
-    selectFileEvents?: EventStoreStatement;
-  };
-};
-
-export type ReplaceFileEventsInput = {
-  source: string;
-  filePath: string;
-  fingerprint: EventStoreFileFingerprint;
-  events: UsageEvent[];
-  skippedRows: number;
-  skippedRowReasons?: SourceSkippedRowReasonStat[];
-  now: number;
-};
-
-export type DeleteStoredFilesInput = {
-  source: string;
-  filePath: string;
-};
-
-export type DeleteStoredFilesResult = {
-  deletedFileCount: number;
-  deletedEventCount: number;
-};
-
-export class EventStoreSchemaVersionError extends Error {
-  readonly schemaVersion: string | undefined;
-
-  constructor(schemaVersion: string | undefined) {
-    const versionLabel = schemaVersion ? `v${schemaVersion}` : 'an unknown schema version';
-    super(
-      `Event store uses ${versionLabel}, which is not supported by this version of llm-usage-metrics (supports v${EVENT_STORE_SCHEMA_VERSION}); leaving store untouched`,
-    );
-    this.name = 'EventStoreSchemaVersionError';
-    this.schemaVersion = schemaVersion;
-  }
-}
-
-function createSchemaSql(): string {
-  return `
-CREATE TABLE IF NOT EXISTS meta (
-  key TEXT PRIMARY KEY,
-  value TEXT NOT NULL
-);
-CREATE TABLE IF NOT EXISTS files (
-  source TEXT NOT NULL,
-  file_path TEXT NOT NULL,
-  fingerprint TEXT NOT NULL,
-  skipped_rows INTEGER NOT NULL,
-  skipped_row_reasons TEXT,
-  ingested_at INTEGER NOT NULL,
-  PRIMARY KEY (source, file_path)
-);
-CREATE TABLE IF NOT EXISTS events (
-  id INTEGER PRIMARY KEY,
-  source TEXT NOT NULL,
-  file_path TEXT NOT NULL,
-  event_index INTEGER NOT NULL,
-  session_id TEXT NOT NULL,
-  timestamp TEXT NOT NULL,
-  model TEXT,
-  provider TEXT,
-  repo_root TEXT,
-  input_tokens INTEGER NOT NULL,
-  output_tokens INTEGER NOT NULL,
-  reasoning_tokens INTEGER NOT NULL,
-  cache_read_tokens INTEGER NOT NULL,
-  cache_write_tokens INTEGER NOT NULL,
-  total_tokens INTEGER NOT NULL,
-  content_hash TEXT,
-  cost_usd REAL,
-  cost_mode TEXT NOT NULL
-);
-CREATE INDEX IF NOT EXISTS events_file ON events(source, file_path, event_index);
-CREATE INDEX IF NOT EXISTS events_content_hash ON events(content_hash);
-`;
-}
-
-function isEventStoreSqliteModule(value: unknown): value is EventStoreSqliteModule {
-  const moduleRecord = asRecord(value);
-  return typeof moduleRecord?.DatabaseSync === 'function';
-}
-
-function normalizeStoreSource(source: string): string {
-  const normalizedSource = normalizeSourceId(source)?.toLowerCase();
-
-  if (!normalizedSource) {
-    throw new Error('Event store source must be a non-empty string');
-  }
-
-  return normalizedSource;
-}
-
-function normalizeStoreFilePath(filePath: string): string {
-  const normalizedFilePath = filePath.trim();
-
-  if (!normalizedFilePath) {
-    throw new Error('Event store file path must be a non-empty string');
-  }
-
-  return normalizedFilePath;
-}
-
-function toNonNegativeInteger(value: unknown): number | undefined {
-  if (typeof value !== 'number' || !Number.isFinite(value) || value < 0) {
-    return undefined;
-  }
-
-  return Math.trunc(value);
-}
-
-function toNonNegativeNumber(value: unknown): number | undefined {
-  if (typeof value !== 'number' || !Number.isFinite(value) || value < 0) {
-    return undefined;
-  }
-
-  return value;
-}
-
-function toText(value: unknown): string | undefined {
-  if (typeof value !== 'string') {
-    return undefined;
-  }
-
-  const normalized = value.trim();
-  return normalized || undefined;
-}
-
-function normalizeDependencyFingerprint(
-  value: EventStoreDependencyFingerprint,
-): EventStoreDependencyFingerprint | undefined {
-  const dependencyPath = toText(value.path);
-
-  if (!dependencyPath) {
-    return undefined;
-  }
-
-  if (!value.exists) {
-    return {
-      path: dependencyPath,
-      exists: false,
-    };
-  }
-
-  const size = toNonNegativeInteger(value.size);
-  const mtimeMs = toNonNegativeNumber(value.mtimeMs);
-
-  if (size === undefined || mtimeMs === undefined) {
-    return undefined;
-  }
-
-  return {
-    path: dependencyPath,
-    exists: true,
-    size,
-    mtimeMs,
-  };
-}
-
-function compareDependencyFingerprint(
-  left: EventStoreDependencyFingerprint,
-  right: EventStoreDependencyFingerprint,
-): number {
-  if (left.path !== right.path) {
-    return compareByCodePoint(left.path, right.path);
-  }
-
-  if (left.exists !== right.exists) {
-    return left.exists ? 1 : -1;
-  }
-
-  if ((left.size ?? -1) !== (right.size ?? -1)) {
-    return (left.size ?? -1) - (right.size ?? -1);
-  }
-
-  return (left.mtimeMs ?? -1) - (right.mtimeMs ?? -1);
-}
-
-function normalizeEventStoreFingerprint(
-  fingerprint: EventStoreFileFingerprint,
-): EventStoreFileFingerprint | undefined {
-  if (!Array.isArray(fingerprint.dependencies) || fingerprint.dependencies.length === 0) {
-    return undefined;
-  }
-
-  const dependencies: EventStoreDependencyFingerprint[] = [];
-
-  for (const dependency of fingerprint.dependencies) {
-    const normalizedDependency = normalizeDependencyFingerprint(dependency);
-
-    if (!normalizedDependency) {
-      return undefined;
-    }
-
-    dependencies.push(normalizedDependency);
-  }
-
-  dependencies.sort(compareDependencyFingerprint);
-
-  return { dependencies };
-}
-
-export function serializeEventStoreFingerprint(fingerprint: EventStoreFileFingerprint): string {
-  const normalizedFingerprint = normalizeEventStoreFingerprint(fingerprint);
-
-  if (!normalizedFingerprint) {
-    throw new Error('Event store fingerprint must include valid dependency fingerprints');
-  }
-
-  return JSON.stringify(normalizedFingerprint);
-}
-
-function parseSkippedRowReasons(value: unknown): SourceSkippedRowReasonStat[] {
-  if (typeof value !== 'string') {
-    return [];
-  }
-
-  try {
-    return normalizeSkippedRowReasons(JSON.parse(value));
-  } catch {
-    return [];
-  }
-}
-
-function stringifySkippedRowReasons(
-  skippedRowReasons: SourceSkippedRowReasonStat[] | undefined,
-): string | null {
-  const normalizedReasons = normalizeSkippedRowReasons(skippedRowReasons);
-  return normalizedReasons.length > 0 ? JSON.stringify(normalizedReasons) : null;
-}
-
-type OptionalStoredText = string | undefined | typeof FAST_PATH_REJECT;
-
-type StoredEventTuple = [
-  source: unknown,
-  session_id: unknown,
-  timestamp: unknown,
-  model: unknown,
-  provider: unknown,
-  repo_root: unknown,
-  input_tokens: unknown,
-  output_tokens: unknown,
-  reasoning_tokens: unknown,
-  cache_read_tokens: unknown,
-  cache_write_tokens: unknown,
-  total_tokens: unknown,
-  cost_usd: unknown,
-  cost_mode: unknown,
-];
-
-export function normalizeStoredEvent(row: Record<string, unknown>): UsageEvent | undefined {
-  return fastMaterializeStoredEvent(row) ?? slowNormalizeStoredEvent(row);
-}
-
-function slowNormalizeStoredEvent(row: Record<string, unknown>): UsageEvent | undefined {
-  const costMode =
-    row.cost_mode === 'explicit' || row.cost_mode === 'estimated' ? row.cost_mode : undefined;
-
-  if (!costMode) {
-    return undefined;
-  }
-
-  const input: UsageEventInput = {
-    source: toText(row.source) ?? '',
-    sessionId: toText(row.session_id) ?? '',
-    timestamp: toText(row.timestamp) ?? '',
-    repoRoot: toText(row.repo_root),
-    provider: toText(row.provider),
-    model: toText(row.model),
-    inputTokens: toNonNegativeInteger(row.input_tokens),
-    outputTokens: toNonNegativeInteger(row.output_tokens),
-    reasoningTokens: toNonNegativeInteger(row.reasoning_tokens),
-    cacheReadTokens: toNonNegativeInteger(row.cache_read_tokens),
-    cacheWriteTokens: toNonNegativeInteger(row.cache_write_tokens),
-    totalTokens: toNonNegativeInteger(row.total_tokens),
-    costUsd: toNonNegativeNumber(row.cost_usd),
-    costMode,
-  };
-
-  try {
-    return createUsageEvent(input);
-  } catch {
-    return undefined;
-  }
-}
-
-function fastMaterializeStoredEvent(row: Record<string, unknown>): UsageEvent | undefined {
-  return materializeStoredEvent(
-    row.source,
-    row.session_id,
-    row.timestamp,
-    row.model,
-    row.provider,
-    row.repo_root,
-    row.input_tokens,
-    row.output_tokens,
-    row.reasoning_tokens,
-    row.cache_read_tokens,
-    row.cache_write_tokens,
-    row.total_tokens,
-    row.cost_usd,
-    row.cost_mode,
-  );
-}
-
-function normalizeStoredEventTuple(row: StoredEventTuple): UsageEvent | undefined {
-  const event = materializeStoredEvent(...row);
-
-  if (event) {
-    return event;
-  }
-
-  return slowNormalizeStoredEvent({
-    source: row[0],
-    session_id: row[1],
-    timestamp: row[2],
-    model: row[3],
-    provider: row[4],
-    repo_root: row[5],
-    input_tokens: row[6],
-    output_tokens: row[7],
-    reasoning_tokens: row[8],
-    cache_read_tokens: row[9],
-    cache_write_tokens: row[10],
-    total_tokens: row[11],
-    cost_usd: row[12],
-    cost_mode: row[13],
-  });
-}
-
-function materializeStoredEvent(
-  sourceValue: unknown,
-  sessionIdValue: unknown,
-  timestampValue: unknown,
-  modelValue: unknown,
-  providerValue: unknown,
-  repoRootValue: unknown,
-  inputTokensValue: unknown,
-  outputTokensValue: unknown,
-  reasoningTokensValue: unknown,
-  cacheReadTokensValue: unknown,
-  cacheWriteTokensValue: unknown,
-  totalTokensValue: unknown,
-  costUsdValue: unknown,
-  costModeValue: unknown,
-): UsageEvent | undefined {
-  const costMode =
-    costModeValue === 'explicit' || costModeValue === 'estimated' ? costModeValue : undefined;
-
-  if (!costMode) {
-    return undefined;
-  }
-
-  const source = toStoredRequiredText(sourceValue);
-  const sessionId = toStoredRequiredText(sessionIdValue);
-  const timestamp = toStoredTimestamp(timestampValue);
-  const repoRoot = toStoredOptionalText(repoRootValue);
-  const provider = toStoredOptionalText(providerValue);
-  const model = toStoredOptionalText(modelValue);
-  const inputTokens = toStoredNonNegativeInteger(inputTokensValue);
-  const outputTokens = toStoredNonNegativeInteger(outputTokensValue);
-  const reasoningTokens = toStoredNonNegativeInteger(reasoningTokensValue);
-  const cacheReadTokens = toStoredNonNegativeInteger(cacheReadTokensValue);
-  const cacheWriteTokens = toStoredNonNegativeInteger(cacheWriteTokensValue);
-  const totalTokens = toStoredNonNegativeInteger(totalTokensValue);
-  const costUsd = toStoredCostUsd(costUsdValue, costMode);
-
-  if (
-    !source ||
-    !sessionId ||
-    !timestamp ||
-    repoRoot === FAST_PATH_REJECT ||
-    provider === FAST_PATH_REJECT ||
-    model === FAST_PATH_REJECT ||
-    inputTokens === undefined ||
-    outputTokens === undefined ||
-    reasoningTokens === undefined ||
-    cacheReadTokens === undefined ||
-    cacheWriteTokens === undefined ||
-    totalTokens === undefined ||
-    costUsd === FAST_PATH_REJECT
-  ) {
-    return undefined;
-  }
-
-  if (provider && normalizeProviderToBillingEntity(provider) !== provider) {
-    return undefined;
-  }
-
-  if (model && model !== model.toLowerCase()) {
-    return undefined;
-  }
-
-  return {
-    source,
-    sessionId,
-    timestamp,
-    repoRoot,
-    provider,
-    model,
-    inputTokens,
-    outputTokens,
-    reasoningTokens,
-    cacheReadTokens,
-    cacheWriteTokens,
-    totalTokens,
-    costUsd,
-    costMode,
-  };
-}
-
-function toStoredRequiredText(value: unknown): string | undefined {
-  if (typeof value !== 'string' || value.length === 0) {
-    return undefined;
-  }
-
-  if (value !== value.trim() || CONTROL_CHARACTERS_PATTERN.test(value)) {
-    return undefined;
-  }
-
-  return value;
-}
-
-function toStoredOptionalText(value: unknown): OptionalStoredText {
-  if (value === null) {
-    return undefined;
-  }
-
-  return toStoredRequiredText(value) ?? FAST_PATH_REJECT;
-}
-
-function toStoredTimestamp(value: unknown): string | undefined {
-  if (typeof value !== 'string' || !NORMALIZED_TIMESTAMP_PATTERN.test(value)) {
-    return undefined;
-  }
-
-  if (!isValidNormalizedTimestamp(value)) {
-    return undefined;
-  }
-
-  return value;
-}
-
-function isValidNormalizedTimestamp(value: string): boolean {
-  const year = Number(value.slice(0, 4));
-  const month = Number(value.slice(5, 7));
-  const day = Number(value.slice(8, 10));
-  const hour = Number(value.slice(11, 13));
-  const minute = Number(value.slice(14, 16));
-  const second = Number(value.slice(17, 19));
-
-  if (month < 1 || month > 12 || day < 1 || hour > 23 || minute > 59 || second > 59) {
-    return false;
-  }
-
-  const maxDay = month === 2 && isLeapYear(year) ? 29 : DAYS_IN_MONTH[month - 1];
-  return day <= maxDay;
-}
-
-function isLeapYear(year: number): boolean {
-  return year % 4 === 0 && (year % 100 !== 0 || year % 400 === 0);
-}
-
-function toStoredNonNegativeInteger(value: unknown): number | undefined {
-  if (
-    typeof value !== 'number' ||
-    !Number.isFinite(value) ||
-    !Number.isInteger(value) ||
-    value < 0
-  ) {
-    return undefined;
-  }
-
-  return value;
-}
-
-function toStoredCostUsd(
-  value: unknown,
-  costMode: 'explicit' | 'estimated',
-): number | undefined | typeof FAST_PATH_REJECT {
-  if (value === null) {
-    return costMode === 'explicit' ? FAST_PATH_REJECT : undefined;
-  }
-
-  if (typeof value !== 'number' || !Number.isFinite(value) || value < 0) {
-    return FAST_PATH_REJECT;
-  }
-
-  return value;
-}
-
-function listUserTableNames(database: EventStoreDatabase): string[] {
-  return database
-    .prepare(
-      [
-        'SELECT name',
-        'FROM sqlite_master',
-        "WHERE type = 'table' AND name NOT LIKE 'sqlite_%'",
-        'ORDER BY name ASC',
-      ].join('\n'),
-    )
-    .all()
-    .map((row) => toText(row.name))
-    .filter((name): name is string => Boolean(name));
-}
-
-function readSchemaVersion(database: EventStoreDatabase): string | undefined {
-  const tableNames = new Set(listUserTableNames(database));
-
-  if (!tableNames.has('meta')) {
-    return undefined;
-  }
-
-  const schemaVersionRow = database
-    .prepare("SELECT value FROM meta WHERE key = 'schemaVersion'")
-    .get();
-  return toText(schemaVersionRow?.value);
-}
-
-function createFreshSchema(database: EventStoreDatabase): void {
-  database.exec(createSchemaSql());
-  database
-    .prepare("INSERT OR REPLACE INTO meta (key, value) VALUES ('schemaVersion', ?)")
-    .run(EVENT_STORE_SCHEMA_VERSION);
-}
-
-export function computeEventContentHash(event: UsageEvent): string {
-  const fields = [
-    event.source.toLowerCase(),
-    event.sessionId,
-    event.timestamp,
-    event.model ?? '',
-    event.provider ?? '',
-    event.repoRoot ?? '',
-    event.inputTokens,
-    event.outputTokens,
-    event.reasoningTokens,
-    event.cacheReadTokens,
-    event.cacheWriteTokens,
-    event.totalTokens,
-    event.costMode,
-    event.costUsd ?? '',
-  ];
-
-  return createHash('sha256').update(fields.map(String).join('\x1f')).digest('hex').slice(0, 16);
-}
-
-export function migrateSchemaV1ToV2(database: EventStoreDatabase): void {
-  runTransaction(database, () => {
-    // Re-check under the write lock: another process may have migrated while
-    // this one waited on BEGIN IMMEDIATE.
-    if (readSchemaVersion(database) !== '1') {
-      return;
-    }
-
-    database.exec('ALTER TABLE events ADD COLUMN content_hash TEXT');
-
-    const selectBatch = database.prepare(
-      [
-        'SELECT id, source, session_id, timestamp, model, provider, repo_root,',
-        '  input_tokens, output_tokens, reasoning_tokens, cache_read_tokens,',
-        '  cache_write_tokens, total_tokens, cost_usd, cost_mode',
-        'FROM events',
-        'WHERE id > ?',
-        'ORDER BY id ASC',
-        'LIMIT ?',
-      ].join('\n'),
-    );
-    const updateContentHash = database.prepare('UPDATE events SET content_hash = ? WHERE id = ?');
-
-    let lastId = -1;
-    for (;;) {
-      const rows = selectBatch.all(lastId, MIGRATION_BATCH_SIZE);
-
-      if (rows.length === 0) {
-        break;
-      }
-
-      for (const row of rows) {
-        const id = toNonNegativeInteger(row.id);
-
-        if (id === undefined) {
-          throw new Error('Cannot migrate event store row without a valid row id');
-        }
-
-        lastId = id;
-
-        const event = normalizeStoredEvent(row);
-
-        // An un-normalizable row keeps a NULL hash instead of aborting the
-        // migration; the history read path never suppresses NULL-hash files.
-        if (!event) {
-          continue;
-        }
-
-        updateContentHash.run(computeEventContentHash(event), id);
-      }
-    }
-
-    database.exec('CREATE INDEX IF NOT EXISTS events_content_hash ON events(content_hash)');
-    database.prepare("UPDATE meta SET value = ? WHERE key = 'schemaVersion'").run('2');
-  });
-}
-
-export function migrateSchemaV2ToV3(database: EventStoreDatabase): void {
-  runTransaction(database, () => {
-    // Re-check under the write lock: another process may have migrated while
-    // this one waited on BEGIN IMMEDIATE.
-    if (readSchemaVersion(database) !== '2') {
-      return;
-    }
-
-    const selectBatch = database.prepare(
-      [
-        'SELECT id, source, session_id, timestamp, model, provider, repo_root,',
-        '  input_tokens, output_tokens, reasoning_tokens, cache_read_tokens,',
-        '  cache_write_tokens, total_tokens, cost_usd, cost_mode',
-        'FROM events',
-        'WHERE id > ?',
-        'ORDER BY id ASC',
-        'LIMIT ?',
-      ].join('\n'),
-    );
-    const updateContentHash = database.prepare('UPDATE events SET content_hash = ? WHERE id = ?');
-
-    let lastId = -1;
-    for (;;) {
-      const rows = selectBatch.all(lastId, MIGRATION_BATCH_SIZE);
-
-      if (rows.length === 0) {
-        break;
-      }
-
-      for (const row of rows) {
-        const id = toNonNegativeInteger(row.id);
-
-        if (id === undefined) {
-          throw new Error('Cannot migrate event store row without a valid row id');
-        }
-
-        lastId = id;
-
-        const event = normalizeStoredEvent(row);
-
-        // An un-normalizable row gets a NULL hash instead of aborting the
-        // migration; the history read path never suppresses NULL-hash files.
-        updateContentHash.run(event ? computeEventContentHash(event) : null, id);
-      }
-    }
-
-    database.prepare("UPDATE meta SET value = ? WHERE key = 'schemaVersion'").run('3');
-  });
-}
-
-// Runs before the WAL pragma so an unsupported store is rejected without
-// persisting a journal-mode change; initializeSchema re-checks under lock.
-function assertSupportedSchemaVersion(database: EventStoreDatabase): void {
-  const schemaVersion = readSchemaVersion(database);
-
-  if (
-    schemaVersion !== undefined &&
-    schemaVersion !== EVENT_STORE_SCHEMA_VERSION &&
-    schemaVersion !== '1' &&
-    schemaVersion !== '2'
-  ) {
-    throw new EventStoreSchemaVersionError(schemaVersion);
-  }
-}
-
-function initializeSchema(database: EventStoreDatabase): void {
-  const tableNames = listUserTableNames(database);
-
-  if (tableNames.length === 0) {
-    createFreshSchema(database);
-    return;
-  }
-
-  const schemaVersion = readSchemaVersion(database);
-
-  if (schemaVersion === EVENT_STORE_SCHEMA_VERSION) {
-    return;
-  }
-
-  if (schemaVersion === '1') {
-    migrateSchemaV1ToV2(database);
-    migrateSchemaV2ToV3(database);
-    return;
-  }
-
-  if (schemaVersion === '2') {
-    migrateSchemaV2ToV3(database);
-    return;
-  }
-
-  throw new EventStoreSchemaVersionError(schemaVersion);
-}
-
-function runTransaction(database: EventStoreDatabase, task: () => void): void {
-  database.exec('BEGIN IMMEDIATE');
-
-  try {
-    task();
-    database.exec('COMMIT');
-  } catch (error) {
-    database.exec('ROLLBACK');
-    throw error;
-  }
-}
-
-function deleteFileEntry(store: EventStore, source: string, filePath: string): void {
-  deleteStoredFiles(store, [{ source, filePath }]);
-}
-
-export function getDefaultEventStorePath(): string {
-  return path.join(getUserCacheRootDir(), 'llm-usage-metrics', 'events.db');
-}
-
-async function prepareEventStoreFile(filePath: string): Promise<void> {
-  const fileHandle = await open(filePath, 'a', 0o600);
-
-  try {
-    await chmod(filePath, 0o600);
-  } finally {
-    await fileHandle.close();
-  }
-}
-
-async function restrictEventStoreFiles(filePath: string): Promise<void> {
-  await chmod(filePath, 0o600);
-
-  for (const sidecarPath of [`${filePath}-wal`, `${filePath}-shm`]) {
-    try {
-      await chmod(sidecarPath, 0o600);
-    } catch (error) {
-      if (
-        typeof error === 'object' &&
-        error !== null &&
-        'code' in error &&
-        error.code === 'ENOENT'
-      ) {
-        continue;
-      }
-
-      throw error;
-    }
-  }
-}
-
-export async function openEventStore(
-  filePath: string = getDefaultEventStorePath(),
-  loadSqliteModule: LoadEventStoreSqliteModule = loadEventStoreSqliteModule,
-): Promise<EventStore> {
-  const parentDirectory = path.dirname(filePath);
-  await mkdir(parentDirectory, { recursive: true, mode: 0o700 });
-
-  if (filePath === getDefaultEventStorePath()) {
-    await chmod(parentDirectory, 0o700);
-  }
-
-  const sqliteModule = await loadSqliteModule();
-
-  if (!isEventStoreSqliteModule(sqliteModule)) {
-    throw new Error('Event store requires a sqlite module with a DatabaseSync constructor');
-  }
-
-  await prepareEventStoreFile(filePath);
-
-  const database = new sqliteModule.DatabaseSync(filePath, {
-    timeout: EVENT_STORE_OPEN_TIMEOUT_MS,
-  });
-
-  try {
-    assertSupportedSchemaVersion(database);
-    database.exec('PRAGMA journal_mode=WAL');
-    initializeSchema(database);
-    await restrictEventStoreFiles(filePath);
-
-    return {
-      database,
-      filePath,
-      statements: {},
-    };
-  } catch (error) {
-    database.close();
-    throw error;
-  }
-}
-
-export type EventStoreSummary = {
-  eventCount: number;
-  schemaVersion?: string;
-};
-
-export type EventStoreStoredFile = {
-  source: string;
-  filePath: string;
-};
-
-export async function readEventStoreSummary(
-  filePath: string = getDefaultEventStorePath(),
-  loadSqliteModule: LoadEventStoreSqliteModule = loadEventStoreSqliteModule,
-): Promise<EventStoreSummary> {
-  const sqliteModule = await loadSqliteModule();
-
-  if (!isEventStoreSqliteModule(sqliteModule)) {
-    throw new Error('Event store requires a sqlite module with a DatabaseSync constructor');
-  }
-
-  const database = new sqliteModule.DatabaseSync(filePath, {
-    readOnly: true,
-    timeout: EVENT_STORE_OPEN_TIMEOUT_MS,
-  });
-
-  try {
-    const schemaVersionRow = database
-      .prepare("SELECT value FROM meta WHERE key = 'schemaVersion'")
-      .get();
-    const countRow = database.prepare('SELECT COUNT(*) AS count FROM events').get();
-
-    return {
-      eventCount: toNonNegativeInteger(countRow?.count) ?? 0,
-      schemaVersion: toText(schemaVersionRow?.value),
-    };
-  } finally {
-    database.close();
-  }
-}
-
-export async function readEventStoreStoredFiles(
-  filePath: string = getDefaultEventStorePath(),
-  loadSqliteModule: LoadEventStoreSqliteModule = loadEventStoreSqliteModule,
-): Promise<EventStoreStoredFile[]> {
-  const sqliteModule = await loadSqliteModule();
-
-  if (!isEventStoreSqliteModule(sqliteModule)) {
-    throw new Error('Event store requires a sqlite module with a DatabaseSync constructor');
-  }
-
-  const database = new sqliteModule.DatabaseSync(filePath, {
-    readOnly: true,
-    timeout: EVENT_STORE_OPEN_TIMEOUT_MS,
-  });
-
-  try {
-    const rows = database
-      .prepare(
-        ['SELECT source, file_path', 'FROM files', 'ORDER BY source ASC, file_path ASC'].join('\n'),
-      )
-      .all();
-    const files: EventStoreStoredFile[] = [];
-
-    for (const row of rows) {
-      const source = toText(row.source);
-      const filePath = toText(row.file_path);
-
-      if (!source || !filePath) {
-        continue;
-      }
-
-      files.push({ source, filePath });
-    }
-
-    return files;
-  } finally {
-    database.close();
-  }
-}
-
-export function getFileEntry(
-  store: EventStore,
-  source: string,
-  filePath: string,
-): EventStoreFileEntry | undefined {
-  const normalizedSource = normalizeStoreSource(source);
-  const normalizedFilePath = normalizeStoreFilePath(filePath);
-  const statement = (store.statements.getFileEntry ??= store.database.prepare(
-    [
-      'SELECT fingerprint, skipped_rows, skipped_row_reasons',
-      'FROM files',
-      'WHERE source = ? AND file_path = ?',
-    ].join('\n'),
-  ));
-  const row = statement.get(normalizedSource, normalizedFilePath);
-
-  if (!row) {
-    return undefined;
-  }
-
-  const fingerprint = toText(row.fingerprint);
-  const skippedRows = toNonNegativeInteger(row.skipped_rows);
-
-  if (!fingerprint || skippedRows === undefined) {
-    deleteFileEntry(store, normalizedSource, normalizedFilePath);
-    return undefined;
-  }
-
-  return {
-    fingerprint,
-    skippedRows,
-    skippedRowReasons: parseSkippedRowReasons(row.skipped_row_reasons),
-  };
-}
-
-function selectFileEventRows(
-  store: EventStore,
-  source: string,
-  filePath: string,
-): StoredEventTuple[] {
-  let statement = store.statements.selectFileEvents;
-
-  if (!statement) {
-    statement = store.database.prepare(
-      [
-        'SELECT source, session_id, timestamp, model, provider, repo_root,',
-        '  input_tokens, output_tokens, reasoning_tokens, cache_read_tokens,',
-        '  cache_write_tokens, total_tokens, cost_usd, cost_mode',
-        'FROM events',
-        'WHERE source = ? AND file_path = ?',
-        'ORDER BY event_index ASC',
-      ].join('\n'),
-    );
-    statement.setReturnArrays(true);
-    store.statements.selectFileEvents = statement;
-  }
-
-  // StatementSync's return type does not narrow after setReturnArrays(true).
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-type-assertion
-  return statement.all(source, filePath) as unknown as StoredEventTuple[];
-}
-
-export function readFileEvents(
-  store: EventStore,
-  source: string,
-  filePath: string,
-): UsageEvent[] | undefined {
-  const normalizedSource = normalizeStoreSource(source);
-  const normalizedFilePath = normalizeStoreFilePath(filePath);
-  const events: UsageEvent[] = [];
-
-  for (const row of selectFileEventRows(store, normalizedSource, normalizedFilePath)) {
-    const event = normalizeStoredEventTuple(row);
-
-    if (!event) {
-      deleteFileEntry(store, normalizedSource, normalizedFilePath);
-      return undefined;
-    }
-
-    events.push(event);
-  }
-
-  return events;
-}
-
-export function readDepartedFileEvents(
-  store: EventStore,
-  source: string,
-  filePath: string,
-): UsageEvent[] {
-  const normalizedSource = normalizeStoreSource(source);
-  const normalizedFilePath = normalizeStoreFilePath(filePath);
-  const events: UsageEvent[] = [];
-
-  for (const row of selectFileEventRows(store, normalizedSource, normalizedFilePath)) {
-    const event = normalizeStoredEventTuple(row);
-
-    // A departed file has no source data left to re-parse, so an invalid row
-    // is skipped instead of deleting the ledger's only copy of the file.
-    if (!event) {
-      continue;
-    }
-
-    events.push(event);
-  }
-
-  return events;
-}
-
-function countMatchingRows(
-  store: EventStore,
-  tableName: 'events' | 'files',
-  source: string,
-  filePath: string,
-): number {
-  const row = store.database
-    .prepare(`SELECT COUNT(*) AS count FROM ${tableName} WHERE source = ? AND file_path = ?`)
-    .get(source, filePath);
-  return toNonNegativeInteger(row?.count) ?? 0;
-}
-
-export function deleteStoredFiles(
-  store: EventStore,
-  files: readonly DeleteStoredFilesInput[],
-): DeleteStoredFilesResult {
-  const normalizedFiles = files.map((file) => ({
-    source: normalizeStoreSource(file.source),
-    filePath: normalizeStoreFilePath(file.filePath),
-  }));
-  const result: DeleteStoredFilesResult = {
-    deletedFileCount: 0,
-    deletedEventCount: 0,
-  };
-
-  runTransaction(store.database, () => {
-    const deleteEvents = store.database.prepare(
-      'DELETE FROM events WHERE source = ? AND file_path = ?',
-    );
-    const deleteFile = store.database.prepare(
-      'DELETE FROM files WHERE source = ? AND file_path = ?',
-    );
-
-    for (const file of normalizedFiles) {
-      result.deletedEventCount += countMatchingRows(store, 'events', file.source, file.filePath);
-      result.deletedFileCount += countMatchingRows(store, 'files', file.source, file.filePath);
-      deleteEvents.run(file.source, file.filePath);
-      deleteFile.run(file.source, file.filePath);
-    }
-  });
-
-  return result;
-}
-
-export function vacuumEventStore(store: EventStore): void {
-  // SQLite requires VACUUM to run outside an explicit transaction.
-  store.database.exec('VACUUM');
-}
-
-export function replaceFileEvents(store: EventStore, input: ReplaceFileEventsInput): void {
-  const source = normalizeStoreSource(input.source);
-  const filePath = normalizeStoreFilePath(input.filePath);
-  const fingerprint = serializeEventStoreFingerprint(input.fingerprint);
-  const skippedRows = toNonNegativeInteger(input.skippedRows) ?? 0;
-  const skippedRowReasons = stringifySkippedRowReasons(input.skippedRowReasons);
-  const ingestedAt = Math.max(0, Math.trunc(input.now));
-  const events = input.events.map((event) => createUsageEvent(event));
-
-  runTransaction(store.database, () => {
-    store.database
-      .prepare('DELETE FROM events WHERE source = ? AND file_path = ?')
-      .run(source, filePath);
-
-    const insertEvent = store.database.prepare(
-      [
-        'INSERT INTO events (',
-        '  source, file_path, event_index, session_id, timestamp, model, provider, repo_root,',
-        '  input_tokens, output_tokens, reasoning_tokens, cache_read_tokens,',
-        '  cache_write_tokens, total_tokens, content_hash, cost_usd, cost_mode',
-        ') VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)',
-      ].join('\n'),
-    );
-
-    events.forEach((event, eventIndex) => {
-      insertEvent.run(
-        source,
-        filePath,
-        eventIndex,
-        event.sessionId,
-        event.timestamp,
-        event.model ?? null,
-        event.provider ?? null,
-        event.repoRoot ?? null,
-        event.inputTokens,
-        event.outputTokens,
-        event.reasoningTokens,
-        event.cacheReadTokens,
-        event.cacheWriteTokens,
-        event.totalTokens,
-        computeEventContentHash(event),
-        event.costUsd ?? null,
-        event.costMode,
-      );
-    });
-
-    store.database
-      .prepare(
-        [
-          'INSERT INTO files (',
-          '  source, file_path, fingerprint, skipped_rows, skipped_row_reasons, ingested_at',
-          ') VALUES (?, ?, ?, ?, ?, ?)',
-          'ON CONFLICT(source, file_path) DO UPDATE SET',
-          '  fingerprint = excluded.fingerprint,',
-          '  skipped_rows = excluded.skipped_rows,',
-          '  skipped_row_reasons = excluded.skipped_row_reasons,',
-          '  ingested_at = excluded.ingested_at',
-        ].join('\n'),
-      )
-      .run(source, filePath, fingerprint, skippedRows, skippedRowReasons, ingestedAt);
-  });
-}
-
-export function countEvents(store: EventStore): number {
-  const row = store.database.prepare('SELECT COUNT(*) AS count FROM events').get();
-  return toNonNegativeInteger(row?.count) ?? 0;
-}
-
-export function closeEventStore(store: EventStore): void {
-  store.database.close();
-}
+export {
+  closeEventStore,
+  countEvents,
+  deleteStoredFiles,
+  getDefaultEventStorePath,
+  getFileEntry,
+  openEventStore,
+  readDepartedFileEvents,
+  readEventStoreStoredFiles,
+  readEventStoreSummary,
+  readFileEvents,
+  replaceFileEvents,
+  serializeEventStoreFingerprint,
+  vacuumEventStore,
+  type DeleteStoredFilesInput,
+  type DeleteStoredFilesResult,
+  type EventStoreDependencyFingerprint,
+  type EventStoreFileEntry,
+  type EventStoreFileFingerprint,
+  type EventStoreStoredFile,
+  type EventStoreSummary,
+  type ReplaceFileEventsInput,
+} from './event-store-administration.js';
+export { computeEventContentHash, normalizeStoredEvent } from './event-store-codec.js';
+export { type EventStore, type LoadEventStoreSqliteModule } from './event-store-database.js';
+export {
+  EVENT_STORE_SCHEMA_VERSION,
+  EventStoreSchemaVersionError,
+  MIGRATION_BATCH_SIZE,
+  migrateSchemaV1ToV2,
+  migrateSchemaV2ToV3,
+} from './event-store-schema.js';
